requirements.txt
# Core Dependencies
python-dateutil>=2.8.2
pytz>=2023.3

# Configuration
PyYAML>=6.0

# Data & Math
numpy>=1.24.0
pandas>=1.5.0
scipy>=1.10.0

# Machine Learning (Vision)
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0

# API & Web
requests>=2.31.0
aiohttp>=3.9.0

# Telegram Bot
python-telegram-bot>=20.0

# Database
sqlalchemy>=2.0.0
sqlite3

# Logging & Monitoring
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Type Checking
mypy>=1.0.0
types-PyYAML>=6.0.0
types-requests>=2.31.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0

# Development
ipython>=8.0.0
jupyter>=1.0.0
